初始化

+++ monitor-console/monitor-web/src/main/java/com/creditease/monitor/controller/MonitorProjectController.java
@@ -57,6 +57,12 @@ public class MonitorProjectController {
             logger.info("deleteProject fail id={} not exists", id);
             return Response.fail(ResponseCode.PROJECT_NOT_EXISTS);
         }
+
+        if(true==monitorMachineService.referProject(monitorProject.getId())){
+            logger.info("deleteProject fail,some machine is referring project id={} ", monitorProject.getId());
+            return Response.fail(ResponseCode.PROJECT_IS_REFERRED_BY_MACHINE_LIST);
+        }
+
         boolean ok = monitorProjectService.deleteProject(monitorProject.getId());
         return Response.ok(ok);
     }
